Refactor: Move PathManager and file modules to new location
Move the PathManager and related file modules from 'utils/file' to 'modules/utils/file'. This change aims to improve the project structure and organization by placing related modules under a more appropriate directory. No functional changes were made to the code.
